refactor(schemas): move users/accounts/sessions to schemas/users/
Per-dialect Drizzle table definitions for the user identity feature group
now live in schemas/users/{postgres,mysql,sqlite}.ts with a dialect-aware
barrel exposing userTables(dialect).

database/schema/<dialect>.ts re-exports from the new location so existing
imports and relations() blocks keep compiling. Relations themselves stay
in database/schema/<dialect>.ts during Plan A and consolidate in Task 17.
