+++ packages/nextly/src/database/schema/postgres.ts
@@ -13,82 +13,18 @@ import {
   type AnyPgColumn,
 } from "drizzle-orm/pg-core";
 
+import { users, accounts, sessions } from "../../schemas/users/postgres";
+
 export const systemMigrations = pgTable("system_migrations", {
   id: serial("id").primaryKey(),
   name: text("name").notNull(),
   runAt: timestamp("run_at", { withTimezone: false }).defaultNow().notNull(),
 });
 
-// Auth.js v5 compatible tables
-export const users = pgTable(
-  "users",
-  {
-    id: text("id").primaryKey(),
-    name: text("name"),
-    email: text("email").notNull(),
-    emailVerified: timestamp("email_verified", { withTimezone: false }),
-    passwordUpdatedAt: timestamp("password_updated_at", {
-      withTimezone: false,
-    }),
-    image: text("image"),
-    passwordHash: text("password_hash").notNull(),
-    isActive: boolean("is_active").notNull().default(false),
-    // Brute-force protection: tracks failed login attempts and account lockout
-    failedLoginAttempts: integer("failed_login_attempts").notNull().default(0),
-    lockedUntil: timestamp("locked_until", { withTimezone: false }),
-    createdAt: timestamp("created_at", { withTimezone: false })
-      .defaultNow()
-      .notNull(),
-    updatedAt: timestamp("updated_at", { withTimezone: false })
-      .defaultNow()
-      .notNull(),
-  },
-  t => [
-    uniqueIndex("users_email_unique").on(t.email),
-    index("users_created_at_idx").on(t.createdAt),
-  ]
-);
-
-export const accounts = pgTable(
-  "accounts",
-  {
-    id: serial("id").primaryKey(),
-    userId: text("user_id").notNull(),
-    type: text("type").notNull(),
-    provider: text("provider").notNull(),
-    providerAccountId: text("provider_account_id").notNull(),
-    refresh_token: text("refresh_token"),
-    access_token: text("access_token"),
-    expires_at: integer("expires_at"),
-    token_type: text("token_type"),
-    scope: text("scope"),
-    id_token: text("id_token"),
-    session_state: text("session_state"),
-    createdAt: timestamp("created_at", { withTimezone: false })
-      .defaultNow()
-      .notNull(),
-    updatedAt: timestamp("updated_at", { withTimezone: false })
-      .defaultNow()
-      .notNull(),
-  },
-  t => [
-    uniqueIndex("accounts_provider_providerAccountId_unique").on(
-      t.provider,
-      t.providerAccountId
-    ),
-    index("accounts_user_id_idx").on(t.userId),
-  ]
-);
-
-export const sessions = pgTable(
-  "sessions",
-  {
-    sessionToken: text("session_token").primaryKey(),
-    userId: text("user_id").notNull(),
-    expires: timestamp("expires", { withTimezone: false }).notNull(),
-  },
-  t => [index("sessions_user_id_idx").on(t.userId)]
-);
+// Auth.js v5 compatible tables — moved to schemas/users/postgres.ts (Plan A Task 5).
+// Re-exported here so existing consumers and relations() blocks below keep working
+// until Task 16 sweeps imports to @nextly/schemas.
+export { users, accounts, sessions } from "../../schemas/users/postgres";
 
 export const verificationTokens = pgTable(
   "verification_tokens",
