+++ packages/nextly/src/database/schema/sqlite.ts
@@ -7,6 +7,8 @@ import {
   uniqueIndex,
 } from "drizzle-orm/sqlite-core";
 
+import { users, accounts, sessions } from "../../schemas/users/sqlite";
+
 export const systemMigrations = sqliteTable("system_migrations", {
   id: integer("id").primaryKey({ autoIncrement: true }),
   name: text("name").notNull(),
@@ -15,70 +17,10 @@ export const systemMigrations = sqliteTable("system_migrations", {
     .$defaultFn(() => new Date()),
 });
 
-// Auth.js v5 compatible tables
-export const users = sqliteTable(
-  "users",
-  {
-    id: text("id").primaryKey(),
-    name: text("name"),
-    email: text("email").notNull(),
-    emailVerified: integer("email_verified", { mode: "timestamp" }),
-    passwordUpdatedAt: integer("password_updated_at", { mode: "timestamp" }),
-    image: text("image"),
-    passwordHash: text("password_hash"),
-    isActive: integer("is_active", { mode: "boolean" })
-      .notNull()
-      .default(false),
-    // Brute-force protection: tracks failed login attempts and account lockout
-    failedLoginAttempts: integer("failed_login_attempts").notNull().default(0),
-    lockedUntil: integer("locked_until", { mode: "timestamp" }),
-    createdAt: integer("created_at", { mode: "timestamp" })
-      .notNull()
-      .$defaultFn(() => new Date()),
-    updatedAt: integer("updated_at", { mode: "timestamp" })
-      .notNull()
-      .$defaultFn(() => new Date()),
-  },
-  t => [
-    uniqueIndex("users_email_unique").on(t.email),
-    index("users_created_at_idx").on(t.createdAt),
-  ]
-);
-
-export const accounts = sqliteTable(
-  "accounts",
-  {
-    id: integer("id").primaryKey({ autoIncrement: true }),
-    userId: text("user_id").notNull(),
-    type: text("type").notNull(),
-    provider: text("provider").notNull(),
-    providerAccountId: text("provider_account_id").notNull(),
-    refresh_token: text("refresh_token"),
-    access_token: text("access_token"),
-    expires_at: integer("expires_at"),
-    token_type: text("token_type"),
-    scope: text("scope"),
-    id_token: text("id_token"),
-    session_state: text("session_state"),
-  },
-  t => [
-    uniqueIndex("accounts_provider_providerAccountId_unique").on(
-      t.provider,
-      t.providerAccountId
-    ),
-    index("accounts_user_id_idx").on(t.userId),
-  ]
-);
-
-export const sessions = sqliteTable(
-  "sessions",
-  {
-    sessionToken: text("session_token").primaryKey(),
-    userId: text("user_id").notNull(),
-    expires: integer("expires", { mode: "timestamp" }).notNull(),
-  },
-  t => [index("sessions_user_id_idx").on(t.userId)]
-);
+// Auth.js v5 compatible tables — moved to schemas/users/sqlite.ts (Plan A Task 5).
+// Re-exported here so existing consumers and relations() blocks below keep working
+// until Task 16 sweeps imports to @nextly/schemas.
+export { users, accounts, sessions } from "../../schemas/users/sqlite";
 
 export const verificationTokens = sqliteTable(
   "verification_tokens",
